Moved data loading and caching in types::Source
Metadata now delegates all data handling to their source member. In
order to implement this, explicit basedir handling has been implemented
in source::Blob.

Blobs with relative pathnames are resolved based on their 'dirname'
member, which however is not serialized when encoding to compact binary:
either an absolute pathname is used, or space is saved by assuming data
is relative to a dataset index, or to a metadata file.

Refactoring is not complete: code compiles, but many tests still fail.

git-svn-id: https://svn.smr.arpa.emr.it/arkimet/arkimet/trunk@2712 9b5092fc-ae44-42f6-b796-0d08dad2e17b